rename AdditionalConsole to SendCommand
