+++ resources/scripts/blueprint/components/Server/Terminal/SendCommand.tsx
